rename sidebar to navigation
