+++ web/src/components/SignInModal.vue
@@ -25,7 +25,7 @@
 </template>
 
 <script lang="ts" setup>
-import { ref, defineProps, defineEmits, watch } from 'vue'
+import { ref, watch } from 'vue'
 import { useRouter } from 'vue-router'
 import { useAuthStore } from '../stores/auth_store'
 import { PATH_PROJECTS } from '../router'
